requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
autopep8==1.6.0
boto3==1.21.15
botocore==1.24.15
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
cssselect2==0.5.0
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-easy-pdf==0.1.1
django-filter==21.1
django-heroku==0.3.1
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
django-storages==1.12.3
djangorestframework==3.13.1
fonttools==4.31.2
future==0.18.2
html5lib==1.1
idna==3.3
importlib-metadata==4.11.1
jmespath==0.10.0
lxml==4.8.0
Markdown==3.3.6
oscrypto==1.3.0
phonenumbers==8.12.45
Pillow==9.0.1
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pydyf==0.1.2
pyHanko==0.12.1
pyhanko-certvalidator==0.19.5
PyPDF3==1.0.6
pyphen==0.12.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
toml==0.10.2
tqdm==4.63.1
tzdata==2022.1
tzlocal==4.1
uritools==4.0.0
urllib3==1.26.8
weasyprint==54.2
webencodings==0.5.1
whitenoise==6.0.0
xhtml2pdf==0.2.7
zipp==3.7.0
zopfli==0.2.1
